actualización guardar id

+++ lib/config/eliminar_usuarios.php
@@ -1,9 +1,15 @@
 <?php 
 include 'setup.php';
+session_start(); 
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     if (isset($_POST['id'])) {
         $id = $_POST['id'];
+
+        if ($id == $_SESSION['user_id']) {
+            echo "No puedes eliminarte a ti mismo";
+            exit;
+        }
         
         $stmt = $con->prepare("DELETE FROM usuarios WHERE idusuarios = ?");
         $stmt->bind_param("i", $id);
